fix some annoying warnings
<==== Mercurial data ====>
Branch: dev_3.0.0

+++ client/libclient/src/plugins/resource/desktop_audio_only/desktop_audio_only_data_provider.cpp
@@ -293,7 +293,7 @@ void QnDesktopAudioOnlyDataProvider::preprocessAudioBuffers(
 
     for( auto& preprocessItem: preprocessList)
     {
-        NX_ASSERT(preprocessItem->buffer.size() >= static_cast<size_t>(kFrameSizeInBytes));
+        NX_ASSERT(preprocessItem->buffer.size() >= static_cast<int>(kFrameSizeInBytes));
 
         auto frameBufferPtr = preprocessItem->frameBuffer;
         memcpy(
